requirements.txt
Django==4.2
mssql-django==1.5
python-decouple==3.8
pyodbc==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
